requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.1
python-dotenv==1.0.1
openai==1.35.10
google-generativeai==0.7.2
duckdb==1.1.3
networkx==3.3